refactor(strategy): 패키지명 수정
