+++ src/main/java/janggi/strategy/ChaStrategy.java
@@ -1,20 +1,18 @@
-package janggi.starategy;
+package janggi.strategy;
 
 import janggi.piece.Piece;
-import janggi.piece.PieceType;
 import janggi.setting.GungSung;
 import janggi.value.Direction;
 import janggi.value.Path;
 import janggi.value.Position;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class PoStrategy implements MoveStrategy {
+public class ChaStrategy implements MoveStrategy {
 
-    private final static List<Direction> DIRECTIONS_OUT_OF_GUNGSUNG =
+    private static List<Direction> DIRECTIONS_OUT_OF_GUNGSUNG =
             List.of(Direction.UP, Direction.DOWN, Direction.LEFT, Direction.RIGHT);
-    private final static List<Direction> DIRECTIONS_IN_OF_GUNGSUNG =
+    private static List<Direction> DIRECTIONS_IN_OF_GUNGSUNG =
             List.of(Direction.UP, Direction.DOWN, Direction.LEFT, Direction.RIGHT,
                     Direction.UP_LEFT, Direction.UP_RIGHT, Direction.DOWN_LEFT, Direction.DOWN_RIGHT);
 
@@ -25,11 +23,10 @@ public class PoStrategy implements MoveStrategy {
             return false;
         }
         Path path = optionalPath.get();
-        List<Piece> piecesInPath = searchPieceInPath(path, enemy, allies);
-        boolean existOnlyOnePieceInPath = existOnlyOnePieceInPath(piecesInPath);
-        boolean existPoInPath = existPoInPath(piecesInPath);
+        boolean existEnemyInPath = existPieceInPath(path, enemy);
+        boolean existAlliesInPath = existPieceInPath(path, allies);
         boolean existAlliesInDestination = existPieceInPosition(destination, allies);
-        return existOnlyOnePieceInPath && !existPoInPath && !existAlliesInDestination;
+        return !existAlliesInPath && !existEnemyInPath && !existAlliesInDestination;
     }
 
     private Optional<Path> calculatePath(Position start, Position destination) {
@@ -42,24 +39,8 @@ public class PoStrategy implements MoveStrategy {
         return Optional.of(new Path(positionsInDirection));
     }
 
-    private List<Piece> searchPieceInPath(Path path, List<Piece> enemy, List<Piece> allies) {
-        ArrayList<Piece> allPieces = new ArrayList<>();
-        allPieces.addAll(enemy);
-        allPieces.addAll(allies);
-        return allPieces.stream()
-                .filter(piece -> path.isInMiddle(piece.getPosition()))
-                .toList();
-    }
-
-    private boolean existOnlyOnePieceInPath(List<Piece> piecesInPath) {
-        return piecesInPath.size() == 1;
-    }
-
-    private boolean existPoInPath(List<Piece> piecesInPath) {
-        long poCount = piecesInPath.stream()
-                .filter(alliesPiece -> alliesPiece.checkPieceType(PieceType.PO))
-                .count();
-        return poCount > 0;
+    private boolean existPieceInPath(Path path, List<Piece> pieces) {
+        return pieces.stream().anyMatch(piece -> path.isInMiddle(piece.getPosition()));
     }
 
     private boolean existPieceInPosition(Position position, List<Piece> pieces) {
